Improve automatic nodeId assignment

+++ src/main/java/de/cas_ual_ty/spells/util/SpellTreeSerializer.java
@@ -62,7 +62,7 @@ public class SpellTreeSerializer
         encodeRequirements(spellNode.getHiddenRequirements(), buf);
         encodeRequirements(spellNode.getLearnRequirements(), buf);
         buf.writeResourceLocation(spellNode.getNodeId().treeId());
-        buf.writeShort(spellNode.getNodeId().nodeId());
+        buf.writeInt(spellNode.getNodeId().nodeId());
         buf.writeByte(spellNode.getFrame());
         buf.writeFloat(spellNode.getSpellInstance().getManaCost());
     }
@@ -115,7 +115,7 @@ public class SpellTreeSerializer
         int levelCost = buf.readInt();
         List<Requirement> hiddenRequirements = decodeRequirements(buf);
         List<Requirement> learnRequirements = decodeRequirements(buf);
-        SpellNodeId id = new SpellNodeId(buf.readResourceLocation(), buf.readShort());
+        SpellNodeId id = new SpellNodeId(buf.readResourceLocation(), buf.readInt());
         int frame = buf.readByte();
         float manaCost = buf.readFloat();
